requirements.txt
streamlit
numpy
pandas
scikit-learn 
Pillow
nltk
plotly
matplotlib
scikit-surprise
fuzzywuzzy
seaborn
st_pages
tmdbv3api